Update view and colorize

+++ github/rebase-fb.sh
@@ -13,22 +13,37 @@ curl -H "Authorization: token ${GIT_TOKEN}" \
     -L "https://api.github.com/repos/exoplatform/swf-jenkins-pipeline/contents/dsl-jobs/FB/seed_jobs_FB_$(echo ${FB_NAME//-} | tr '[:upper:]' '[:lower:]').groovy" --output fblist.txt
 cat fblist.txt | grep "project:" > fblistfiltred.txt
 
+### Colors
+red=`tput setaf 1`
+green=`tput setaf 2`
+magenta=`tput setaf 2`
+reset=`tput sgr0`
+###
+
 echo "Done. Performing action..."
 while IFS= read -r line; do
     item=$(echo $line | awk -F'project:' '{print $2}' | cut -d "," -f 1 | tr -d "'"| xargs)
     org=$(echo $line | awk -F'gitOrganization:' '{print $2}' | cut -d "," -f 1 | tr -d "'" | tr -d "]"| xargs)
     [ -z "${item}" ] && continue
     [ -z "${org}" ] && continue
-    echo "Rebasing ${org}/${item}:feature/${FB_NAME}..."
+    echo "================================================================================================="
+    echo " Module: ${green}${org}/${item}${reset}"
+    echo "================================================================================================="
     git init $item &>/dev/null
     pushd $item &>/dev/null
-    git remote add -t develop -t feature/${FB_NAME} origin git@github.com:${org}/${item}.git
+    git remote add -t develop -t feature/${FB_NAME} origin git@github.com:${org}/${item}.git &>/dev/null
     default_branch=$(git remote show origin | grep 'HEAD branch' | cut -d' ' -f5)
     git fetch &>/dev/null
     default_branch="develop"
-    git checkout feature/${FB_NAME} 
+    git checkout feature/${FB_NAME} &>/dev/null
     git rebase origin/$default_branch feature/${FB_NAME}
     git log --oneline --cherry origin/$default_branch..HEAD
-    git push origin feature/${FB_NAME} --force-with-lease
+    if [ "$(git log HEAD..FETCH_HEAD --oneline 2>/dev/null | wc -l)" -gt "0" ]; then
+      echo "${magenta}Changes before the rebase:{reset}"
+      git log HEAD..FETCH_HEAD --oneline --format="(%h) %s"
+      git push origin feature/${FB_NAME} --force-with-lease
+    else 
+      echo "${green}INFO:${reset} No changes detected!"  
+    fi
     popd &>/dev/null
 done < fblistfiltred.txt
